requirements.txt
Flask==3.1.3
python-dotenv==1.0.1
google-api-python-client==2.170.0
google-auth==2.40.3
reportlab==4.2.2
